Challenge 9.5
I haven't done the whole challenge, just the rocket taking off because it took me long enough to do that on its own. I have made the camera look around at the rocket before it takes off using dolly cams and different blends, and while taking off. I've made use of a canvas animation in order to fade out and in between different cameras so the cuts aren't so abrupt.